[asm] allow changes to selected contests after audit review
[#158824218]
