+++ client/src/component/DOS/DefineAudit/ReviewPage.tsx
@@ -1,8 +1,10 @@
 import * as React from 'react';
 
+import * as _ from 'lodash';
+
 import Nav from '../Nav';
 
-import {SelectedContests} from './SelectContestsPage';
+import counties from 'corla/data/counties';
 
 const Breadcrumb = () => (
     <ul className='pt-breadcrumbs'>
@@ -23,6 +25,56 @@ const Breadcrumb = () => (
         </li>
     </ul>
 );
+
+function formatReason(reason: AuditReason): string {
+    if (reason === 'STATE_WIDE_CONTEST') {
+        return 'State Contest';
+    }
+
+    return 'County Contest';
+}
+
+interface SelectedContestsProps {
+    auditedContests: DOS.AuditedContests;
+    contests: DOS.Contests;
+}
+
+const SelectedContests = (props: SelectedContestsProps) => {
+    const { auditedContests, contests } = props;
+
+    const rows = _.map(props.auditedContests, audited => {
+        const contest = contests[audited.id];
+        const countyName = counties[contest.countyId].name;
+
+        return (
+            <tr key={ contest.id }>
+                <td>{ countyName }</td>
+                <td>{ contest.name }</td>
+                <td>{ formatReason(audited.reason) }</td>
+            </tr>
+        );
+    });
+
+    return (
+        <div className='pt-card'>
+            <h3>Selected Contests</h3>
+            <div className='pt-card'>
+                <table className='pt-table pt-bordered pt-condensed'>
+                    <thead>
+                        <tr>
+                            <th>County</th>
+                            <th>Name</th>
+                            <th>Reason</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        { rows }
+                    </tbody>
+                </table>
+            </div>
+        </div>
+    );
+};
 
 interface AuditReviewProps {
     back: OnClick;
