+++ client/src/component/DOS/DefineAudit/SelectContestsPage.tsx
@@ -6,8 +6,6 @@ import Nav from '../Nav';
 
 import SelectContestsForm from './SelectContestsForm';
 
-import counties from 'corla/data/counties';
-
 
 const Breadcrumb = () => (
     <ul className='pt-breadcrumbs'>
@@ -28,56 +26,6 @@ const Breadcrumb = () => (
         </li>
     </ul>
 );
-
-function formatReason(reason: AuditReason): string {
-    if (reason === 'STATE_WIDE_CONTEST') {
-        return 'State Contest';
-    }
-
-    return 'County Contest';
-}
-
-interface SelectedContestsProps {
-    auditedContests: DOS.AuditedContests;
-    contests: DOS.Contests;
-}
-
-export const SelectedContests = (props: SelectedContestsProps) => {
-    const { auditedContests, contests } = props;
-
-    const rows = _.map(props.auditedContests, audited => {
-        const contest = contests[audited.id];
-        const countyName = counties[contest.countyId].name;
-
-        return (
-            <tr key={ contest.id }>
-                <td>{ countyName }</td>
-                <td>{ contest.name }</td>
-                <td>{ formatReason(audited.reason) }</td>
-            </tr>
-        );
-    });
-
-    return (
-        <div className='pt-card'>
-            <h3>Selected Contests</h3>
-            <div className='pt-card'>
-                <table className='pt-table pt-bordered pt-condensed'>
-                    <thead>
-                        <tr>
-                            <th>County</th>
-                            <th>Name</th>
-                            <th>Reason</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        { rows }
-                    </tbody>
-                </table>
-            </div>
-        </div>
-    );
-};
 
 interface WaitingPageProps {
     back: OnClick;
@@ -131,24 +79,19 @@ const SelectContestsPage = (props: PageProps) => {
     const haveSelectedContests = !_.isEmpty(auditedContests);
 
     const onSaveAndNext = () => {
-        if (!haveSelectedContests) {
-            selectContestsForAudit(forms.selectContestsForm);
-        }
+        selectContestsForAudit(forms.selectContestsForm);
         nextPage();
     };
-
-    const contentDiv = !haveSelectedContests
-                     ? <SelectContestsForm forms={ forms }
-                                           contests={ contests }
-                                           isAuditable={ isAuditable } />
-                     : <SelectedContests auditedContests={ auditedContests }
-                                         contests={ contests } />;
 
     return (
         <div>
             <Nav />
             <Breadcrumb />
-            { contentDiv }
+            <SelectContestsForm forms={ forms }
+                                contests={ contests }
+                                auditedContests={auditedContests}
+                                isAuditable={ isAuditable } />
+
             <div>
                 <button onClick={ back } className='pt-button'>
                     Back
