+++ server/eclipse-project/src/main/java/us/freeandfair/corla/asm/ASMTransitionFunction.java
@@ -49,7 +49,16 @@ public interface ASMTransitionFunction {
     D(new ASMTransition(RANDOM_SEED_PUBLISHED, 
                         DOS_START_ROUND_EVENT,
                         DOS_AUDIT_ONGOING)),
-    E(new ASMTransition(DOS_AUDIT_ONGOING, 
+    D2(new ASMTransition(RANDOM_SEED_PUBLISHED,
+                         COMPLETE_AUDIT_INFO_EVENT,
+                         COMPLETE_AUDIT_INFO_SET)),
+    D3(new ASMTransition(RANDOM_SEED_PUBLISHED,
+                         PARTIAL_AUDIT_INFO_EVENT,
+                         PARTIAL_AUDIT_INFO_SET)),
+    D4(new ASMTransition(COMPLETE_AUDIT_INFO_SET,
+                         DOS_START_ROUND_EVENT,
+                         DOS_AUDIT_ONGOING)),
+    E(new ASMTransition(DOS_AUDIT_ONGOING,
                         SetCreator.setOf(AUDIT_EVENT,
                                          DOS_COUNTY_AUDIT_COMPLETE_EVENT,
                                          DOS_START_ROUND_EVENT),
