+++ server/eclipse-project/src/main/java/us/freeandfair/corla/endpoint/SelectContestsForAudit.java
@@ -89,13 +89,15 @@ public class SelectContestsForAudit extends AbstractDoSDashboardEndpoint {
         Main.LOGGER.error("could not get department of state dashboard");
         serverError(the_response, "Could not select contests");
       } else {
+        // unchecked contests are not posted so that which is not added, is removed.
+        dosdb.removeContestsToAudit();
         for (final ContestToAudit c : contests) {
           Main.LOGGER.info("updating contest audit status: " + c);
           dosdb.updateContestToAudit(c);
           Persistence.saveOrUpdate(dosdb);
-          my_event.set(nextEvent(dosdb));
-          ok(the_response, "Contests selected");
         }
+        my_event.set(nextEvent(dosdb));
+        ok(the_response, "Contests selected");
       }
     } catch (final JsonParseException e) {
       Main.LOGGER.error("malformed contest selection");
